# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Maiko API Documentation",
        "contact": {
            "email": "juancho637@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/maiko/public/api/v1"
        }
    ],
    "paths": {
        "/answers/{answer}": {
            "get": {
                "tags": [
                    "Respuestas"
                ],
                "summary": "Muestra la información de una respuesta",
                "operationId": "App\\Http\\Controllers\\Api\\Answer\\AnswerController::show",
                "parameters": [
                    {
                        "name": "answer",
                        "in": "path",
                        "description": "Id de la respuesta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra la información asociada a una respuesta.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Id no encontrado."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Respuestas"
                ],
                "summary": "Actualiza una respuesta",
                "operationId": "App\\Http\\Controllers\\Api\\Answer\\AnswerController::update",
                "parameters": [
                    {
                        "name": "answer",
                        "in": "path",
                        "description": "Id de la respuesta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "value": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualiza una respuesta.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/companies/{id}/clients": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listado de clientes",
                "operationId": "App\\Http\\Controllers\\Api\\Company\\CompanyClientController::index",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra el listado de clientes asociadas a una empresa.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Id no encontrado."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Almacena un cliente",
                "operationId": "App\\Http\\Controllers\\Api\\Company\\CompanyClientController::store",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la empresa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Almacena un nuevo cliente asociado a una empresa.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}